fix bug sidebar
side bar aktif

+++ resources/views/kos/listKos.blade.php
@@ -11,12 +11,7 @@
                 </div>
                 <div class="d-flex">
                     <form action="/cari-kos" method="get" class="d-flex me-1">
-                        <input
-                            type="text"
-                            name="cari"
-                            id="cari"
-                            class="form-control me-1"
-                        />
+                        <input type="text" name="cariKos" id="cariKos" class="form-control me-1" />
                         <button class="btn btn-primary">
                             <i class="bi bi-search"></i>
                         </button>
@@ -24,23 +19,23 @@
                 </div>
             </div>
             <!-- flash message insert -->
-            @if(Session::has('insert'))
+            @if (Session::has('insert'))
             <div class="alert alert-success mt-3">
-                {{ Session::get('pesan')}}
+                {{ Session::get('pesan') }}
             </div>
             @endif
 
             <!-- flash message update -->
-            @if(Session::has('update'))
+            @if (Session::has('update'))
             <div class="alert alert-success mt-3">
-                {{ Session::get('pesan')}}
+                {{ Session::get('pesan') }}
             </div>
             @endif
 
             <!-- flash message delete -->
-            @if(Session::has('delete'))
+            @if (Session::has('delete'))
             <div class="alert alert-danger mt-3">
-                {{ Session::get('pesan')}}
+                {{ Session::get('pesan') }}
             </div>
             @endif
         </div>
@@ -58,23 +53,17 @@
             <tbody class="">
                 <tr class="table-striped">
                     <td>{{ $loop->iteration }}</td>
-                    <td>{{ $k->nama}}</td>
-                    <td>{{ $k->alamat}}</td>
-                    <td>{{ $k->no_telp}}</td>
+                    <td>{{ $k->nama }}</td>
+                    <td>{{ $k->alamat }}</td>
+                    <td>{{ $k->no_telp }}</td>
                     <td style="width: 100px; white-space: nowrap">
                         <a href="/kamar/{{ $k->id }}" class="btn btn-warning">
-                            <i class="bi bi-info-circle-fill"></i
-                        ></a>
-                        <a href="/edit-kos/{{$k->id}}" class="btn btn-success">
-                            <i class="bi bi-pencil-fill"></i
-                        ></a>
-                        <a
-                            href="/delete-kos/{{$k->id}}"
-                            onclick="return confirm('Hapus data {{ $k->nama }} ?')"
-                            class="btn btn-danger"
-                        >
-                            <i class="bi bi-trash-fill"></i
-                        ></a>
+                            <i class="bi bi-info-circle-fill"></i></a>
+                        <a href="/edit-kos/{{ $k->id }}" class="btn btn-success">
+                            <i class="bi bi-pencil-fill"></i></a>
+                        <a href="/delete-kos/{{ $k->id }}" onclick="return confirm('Hapus data {{ $k->nama }} ?')"
+                            class="btn btn-danger">
+                            <i class="bi bi-trash-fill"></i></a>
                     </td>
                 </tr>
             </tbody>
